Fix toHaveStyle not to return true when shouldn't
[As the docs says](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/style), the `style` property of an HTMLElement should not be directly assigned:
```
Styles should not be set by assigning a string directly to the style property (as in elt.style = "color: blue;"), since it is considered read-only, as the style attribute returns a CSSStyleDeclaration object which is also read-only. Instead, styles can be set by assigning values to the properties of style.
```
By doing that, whenever a nonexistent style was passed to `getStyleDeclaration`, it would return an empty object, making `toHaveStyle` always return true in these cases.

Since the goal was to obtain a key value pair of style properties, `getStyleDeclaration` was changed to parse the string and obtain an object.
That object could be returned, but the HTMLElement style attribution was kept to normalize colors.