Handle extend-expect.js
This was previously handled as a file in the root of the project that redirected to the built index file.
This change removes a layer of indirection by removing src/extend-expect, and using
exports to point to the index file rather than using a real file in the project root.